+++ docs/AngularCache.html
@@ -254,7 +254,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 666
+        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 673
     </li></ul></dd>
     
     
@@ -411,7 +411,7 @@ requested item has expired.</td>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 562
+        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 567
     </li></ul></dd>
     
     
@@ -498,7 +498,7 @@ requested item has expired.</td>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 689
+        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 696
     </li></ul></dd>
     
     
@@ -585,7 +585,7 @@ requested item has expired.</td>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 709
+        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 729
     </li></ul></dd>
     
     
@@ -672,7 +672,7 @@ requested item has expired.</td>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 699
+        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 719
     </li></ul></dd>
     
     
@@ -850,7 +850,7 @@ requested item has expired.</td>
             
             
             
-            <td class="description last">{ maxAge: {Number} }</td>
+            <td class="description last">{{ maxAge: {Number}, aggressiveDelete: {Boolean}, timestamp: {Number} }}</td>
         </tr>
 	
 	
@@ -881,7 +881,7 @@ requested item has expired.</td>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 492
+        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 496
     </li></ul></dd>
     
     
@@ -1016,7 +1016,7 @@ requested item has expired.</td>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 609
+        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 616
     </li></ul></dd>
     
     
@@ -1081,7 +1081,7 @@ requested item has expired.</td>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 642
+        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 649
     </li></ul></dd>
     
     
@@ -1218,7 +1218,7 @@ applying the new options, otherwise only the options specified in the hash will 
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 719
+        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 739
     </li></ul></dd>
     
     
@@ -1261,7 +1261,7 @@ applying the new options, otherwise only the options specified in the hash will 
 <br clear="both">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.2.0-dev</a> on Fri Sep 20 2013 20:27:45 GMT-0600 (MDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.2.0-dev</a> on Sat Sep 21 2013 00:17:40 GMT-0600 (MDT)
 </footer>
 
 <script> prettyPrint(); </script>
