+++ docs/AngularCacheFactory.html
@@ -167,7 +167,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 733
+        <a href="angular-cache-1.2.0-SNAPSHOT.js.html">angular-cache-1.2.0-SNAPSHOT.js</a>, line 753
     </li></ul></dd>
     
     
@@ -242,7 +242,7 @@
 <br clear="both">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.2.0-dev</a> on Fri Sep 20 2013 20:27:45 GMT-0600 (MDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.2.0-dev</a> on Sat Sep 21 2013 00:17:40 GMT-0600 (MDT)
 </footer>
 
 <script> prettyPrint(); </script>
