+++ docs/angular-cache-1.2.0-SNAPSHOT.js.html
@@ -325,7 +325,7 @@
                                     if (!cacheDirty) {
                                         _loadCacheConfig();
                                     } else {
-                                        _saveCacheConfig();
+                                        _syncToStorage(null);
                                         keys = _keys(data);
                                         for (i = 0; i &lt; keys.length; i++) {
                                             storage.setItem(prefix + '.data.' + keys[i], angular.toJson(data[keys[i]]));
@@ -340,7 +340,7 @@
                                     if (!cacheDirty) {
                                         _loadCacheConfig();
                                     } else {
-                                        _saveCacheConfig();
+                                        _syncToStorage(null);
                                         keys = _keys(data);
                                         for (i = 0; i &lt; keys.length; i++) {
                                             storage.setItem(prefix + '.data.' + keys[i], angular.toJson(data[keys[i]]));
@@ -497,22 +497,26 @@
                                 if (data.hasOwnProperty('aggressiveDelete')) {
                                     options.aggressiveDelete = data.aggressiveDelete;
                                 }
-                                self.put(keys[i], data.value);
+                                self.put(keys[i], data.value, options);
                             }
                         }
-                        _saveCacheConfig();
+                        _syncToStorage(null, null);
                     }
                 }
 
                 /**
-                 * @method _saveCacheConfig
-                 * @desc If storageMode is set, save current keys of cache to localStorage.
+                 * @method _syncToStorage
+                 * @desc If storageMode is set, sync to localStorage.
+                 * @param {String} key The identifier of the item to sync.
                  * @private
                  * @ignore
                  */
-                function _saveCacheConfig() {
+                function _syncToStorage(key) {
                     if (config.storageMode && storage) {
                         storage.setItem(prefix + '.keys', angular.toJson(_keys(data)));
+                        if (key) {
+                            storage.setItem(prefix + '.data.' + key, angular.toJson(data[key]));
+                        }
                     }
                 }
 
@@ -521,7 +525,7 @@
                  * @desc Add a key-value pair with timestamp to the cache.
                  * @param {String} key The identifier for the item to add to the cache.
                  * @param {*} value The value of the item to add to the cache.
-                 * @param {Object} [options] { maxAge: {Number} }
+                 * @param {Object} [options] {{ maxAge: {Number}, aggressiveDelete: {Boolean}, timestamp: {Number} }}
                  * @returns {*} value The value of the item added to the cache.
                  * @privileged
                  */
@@ -552,11 +556,19 @@
 
                     if (!(key in data)) {
                         size++;
+                    } else {
+                        if (data[key].timeoutId) {
+                            $timeout.cancel(data[key].timeoutId);
+                        }
                     }
 
                     data[key] = {
                         value: value
                     };
+
+                    if (options && options.maxAge) {
+                        data[key].maxAge = options.maxAge;
+                    }
 
                     if (options && options.hasOwnProperty('aggressiveDelete')) {
                         data[key].aggressiveDelete = options.aggressiveDelete;
@@ -565,19 +577,12 @@
                     data[key].timestamp = (options && options.timestamp) || new Date().getTime();
 
                     if ((options && options.maxAge) || config.maxAge) {
-                        if (data[key].timeoutId) {
-                            $timeout.cancel(data[key].timeoutId);
-                        }
                         if (data[key].aggressiveDelete || (!data[key].hasOwnProperty('aggressiveDelete') && config.aggressiveDelete)) {
                             _setTimeoutToRemove(key, ((options && options.maxAge) || config.maxAge));
                         }
                     }
 
-                    _saveCacheConfig();
-
-                    if (config.storageMode) {
-                        storage.setItem(prefix + '.data.' + key, angular.toJson(data[key]));
-                    }
+                    _syncToStorage(key);
 
                     if (size > config.capacity) {
                         this.remove(staleEnd.key);
@@ -630,6 +635,8 @@
 
                     _refresh(lruEntry);
 
+                    _syncToStorage(key);
+
                     return item.value;
                 };
 
@@ -657,7 +664,7 @@
                     delete lruHash[key];
                     delete data[key];
 
-                    _saveCacheConfig();
+                    _syncToStorage(null);
 
                     if (config.storageMode) {
                         storage.removeItem(prefix + '.data.' + key);
@@ -686,7 +693,7 @@
                     staleEnd = null;
 
                     if (config.storageMode) {
-                        _saveCacheConfig();
+                        _syncToStorage();
                     }
                 };
 
@@ -719,8 +726,21 @@
                  * @returns {Object} stats Object containing information about this cache.
                  * @privileged
                  */
-                this.info = function () {
-                    return angular.extend({}, config, { size: size });
+                this.info = function (key) {
+                    if (key in data) {
+                        var info = {
+                            timestamp: data[key].timestamp,
+                            maxAge: data[key].maxAge || config.maxAge,
+                            aggressiveDelete: data[key].aggressiveDelete || (!data[key].hasOwnProperty('aggressiveDelete') && config.aggressiveDelete) || false,
+                            isExpired: false
+                        };
+                        if (info.maxAge) {
+                            info.isExpired = (new Date().getTime() - info.timestamp) > info.maxAge;
+                        }
+                        return info;
+                    } else {
+                        return angular.extend({}, config, { size: size });
+                    }
                 };
 
                 /**
@@ -872,7 +892,7 @@
 <br clear="both">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.2.0-dev</a> on Fri Sep 20 2013 20:27:45 GMT-0600 (MDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.2.0-dev</a> on Sat Sep 21 2013 00:17:40 GMT-0600 (MDT)
 </footer>
 
 <script> prettyPrint(); </script>
