+++ docs/index.html
@@ -154,7 +154,7 @@ $cacheFactory provided by AngularJS.</div>
 <br clear="both">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.2.0-dev</a> on Fri Sep 20 2013 20:27:45 GMT-0600 (MDT)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.2.0-dev</a> on Sat Sep 21 2013 00:17:40 GMT-0600 (MDT)
 </footer>
 
 <script> prettyPrint(); </script>
